voice_db_clean/requirements.txt
fastapi
uvicorn
torch
torchaudio
speechbrain
librosa
soundfile
faiss-cpu
numpy
python-multipart
qdrant-client
python-dotenv
SpeechRecognition
